% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{makeApp}
\alias{makeApp}
\title{Create a Shiny app object from an R function}
\usage{
makeApp(fun, withGoButton = FALSE)
}
\arguments{
\item{fun}{A function 
(preferably a symbol -- a long self-explanatory name -- pointing to a pre-defined
function object, rather than an anonymous function) 
with zero or more arguments/parameters. \strong{Every argument must have a default value},
which will be used to define each argument's:
\itemize{
  \item type/class,
  \item allowed values,
  \item pre-selected/start-up value.
}}

\item{withGoButton}{Either \code{TRUE} or \code{FALSE} (default: \code{FALSE}).
It indicates if the (re)evaluation of \code{fun} in the Shiny app should be immediately
triggered  by every change in the value of any argument/parameter
(\code{withGoButton} = \code{FALSE}) or if the (re)calculation should be started
only when a specific button is pressed (\code{withGoButton} = \code{TRUE}).
The latter is preferred if the (re)evaluation of \code{fun} is significantly
time-consuming or if \code{fun} has no arguments (because then, without the button,
only refreshing the web page would trigger the (re)evaluation).}
}
\value{
A Shiny app object as returned by \code{\link[shiny]{as.shiny.appobj}}.
}
\description{
Create a Shiny app object from an R function
}
