% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{makeFiles}
\alias{makeFiles}
\title{Create Shiny app files from an R function}
\usage{
makeFiles(fun, withGoButton = FALSE)
}
\value{
\code{NULL}. This function saves two plain text files:
\code{ui.R} and \code{server.R} with the R code of function \code{fun}
translated into a Shiny app. If these files need further manual changes,
it is recommended that they are first re-formatted e.g. in RStudio
(top menu -> Code -> Reformat Code or Ctrl+Shift+A) or programmatically
(e.g. \url{https://github.com/google/rfmt}).
}
\description{
For the description of arguments/parameters, see the help file for
\code{\link{makeApp}}.
}
